Extract `UIRefreshControl` creation/configuration and refresh logic with `FeedLoader` to the new `FeedRefreshViewController`
